Extract CLI stub fixture
